requirements.txt
numpy>= 2.0.2
scikit-learn>=1.6.1
PyWavelets>=1.6.0
opencv-python>=4.5.0
seaborn>=0.12.0
Flask>=3.1.2
